refactor: Header 내 CartQuantity 컴포넌트 분리
