+++ src/component/@common/CardQuantity/CardQuantity.styles.ts
@@ -1,0 +1,18 @@
+import { css } from '@emotion/react';
+
+export const CartQuantityStyle = css`
+  position: absolute;
+  top: 70%;
+  right: 0;
+  transform: translateY(-50%);
+  width: 1.9rem;
+  height: 1.9rem;
+  border-radius: 50%;
+  background-color: #fff;
+  color: #000;
+  font-size: 1rem;
+  font-weight: 700;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
